fix: forgot to stage some changes caused by the refactor
