+++ src/main/java/tech/gruppone/stalker/server/controllers/OrganizationController.java
@@ -2,8 +2,6 @@ package tech.gruppone.stalker.server.controllers;
 
 import java.io.IOException;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-
 import org.springframework.boot.autoconfigure.security.SecurityProperties.User;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -16,7 +14,7 @@ import org.springframework.web.bind.annotation.RestController;
 import lombok.Value;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
-import tech.gruppone.stalker.server.models.api.Organization;
+import tech.gruppone.stalker.server.model.api.Organization;
 import tech.gruppone.stalker.server.repositories.OrganizationRepository;
 
 @RequestMapping("/organization/{id}")
