+++ src/main/java/tech/gruppone/stalker/server/controllers/OrganizationsController.java
@@ -2,8 +2,6 @@ package tech.gruppone.stalker.server.controllers;
 
 import java.io.IOException;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -13,7 +11,7 @@ import org.springframework.web.bind.annotation.RestController;
 import lombok.Value;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
-import tech.gruppone.stalker.server.models.api.Organization;
+import tech.gruppone.stalker.server.model.api.Organization;
 import tech.gruppone.stalker.server.repositories.OrganizationRepository;
 
 @RequestMapping("/organizations")
