+++ src/main/java/tech/gruppone/stalker/server/model/api/OrganizationData.java
@@ -1,4 +1,4 @@
-package tech.gruppone.stalker.server.models.api;
+package tech.gruppone.stalker.server.model.api;
 
 import java.sql.Timestamp;
 import java.util.ArrayList;
