+++ src/main/java/tech/gruppone/stalker/server/model/api/PlaceData.java
@@ -1,4 +1,4 @@
-package tech.gruppone.stalker.server.models.api;
+package tech.gruppone.stalker.server.model.api;
 
 import java.util.List;
 import lombok.Data;
