+++ src/main/java/tech/gruppone/stalker/server/model/api/TimeInsidePlaces.java
@@ -1,4 +1,4 @@
-package tech.gruppone.stalker.server.models.api;
+package tech.gruppone.stalker.server.model.api;
 
 import java.util.List;
