+++ src/main/java/tech/gruppone/stalker/server/model/api/UnauthenticatedUser.java
@@ -1,4 +1,4 @@
-package tech.gruppone.stalker.server.models.api;
+package tech.gruppone.stalker.server.model.api;
 
 import lombok.Data;
 import lombok.NonNull;
