+++ src/main/java/tech/gruppone/stalker/server/model/api/requests/PutUserIdRequest.java
@@ -1,10 +1,10 @@
-package tech.gruppone.stalker.server.models.api.request;
+package tech.gruppone.stalker.server.model.api.requests;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.NonNull;
-import tech.gruppone.stalker.server.models.api.LoginData;
-import tech.gruppone.stalker.server.models.api.UserData;
+import tech.gruppone.stalker.server.model.api.LoginData;
+import tech.gruppone.stalker.server.model.api.UserData;
 
 @Data
 @NoArgsConstructor
